Ui and script updates
Added button tabs for inventory, skills menu, and upgrade menu to freely switch between. Made a script to enable other scripts, this was done to run the name input scene script when the new game button was pressed. Fixed an issue with the options menu not saving the settings by removing the back button and only having the save button present.